% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatA_Palette.R
\name{display_palette}
\alias{display_palette}
\title{Display and return a color palette}
\usage{
display_palette(palette_name)
}
\arguments{
\item{palette_name}{The name of the color palette to display (e.g., "yellow6").}
}
\value{
A vector of color hex codes.
}
\description{
This function returns the specified color palette and also displays it using a barplot.
}
\examples{
display_palette("yellow6")
}
