% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatA_Palette.R
\name{get_palette}
\alias{get_palette}
\title{Get Palette}
\usage{
get_palette(name)
}
\arguments{
\item{name}{A string, name of the palette to retrieve.}
}
\value{
A vector of color hex codes.
}
\description{
Retrieve a color palette by name.
}
\examples{
get_palette("brown1")
get_palette("brown2")
get_palette("brown3")
get_palette("brown4")
get_palette("brown5")
get_palette("brown6")
get_palette("blue1")
get_palette("blue2")
get_palette("blue3")
get_palette("blue4")
get_palette("blue5")
get_palette("blue6")
get_palette("red1")
get_palette("red2")
get_palette("red3")
get_palette("red4")
get_palette("red5")
get_palette("red6")
get_palette("violet1")
get_palette("violet2")
get_palette("violet3")
get_palette("violet4")
get_palette("violet5")
get_palette("violet6")
get_palette("yellow1")
get_palette("yellow2")
get_palette("yellow3")
get_palette("yellow4")
get_palette("yellow5")
get_palette("yellow6")
get_palette("azure1")
get_palette("azure2")
get_palette("azure3")
get_palette("azure4")
get_palette("azure5")
get_palette("azure6")
get_palette("green1")
get_palette("green2")
get_palette("green3")
get_palette("green4")
get_palette("green5")
get_palette("green6")
get_palette("rainbow1")
get_palette("rainbow2")
get_palette("rainbow3")
get_palette("rainbow4")
get_palette("rainbow5")
get_palette("pastel1")
get_palette("pastel2")
get_palette("pastel3")
get_palette("pastel4")
get_palette("pastel5")
get_palette("pastel6")
get_palette("pastel7")
get_palette("pastel8")
get_palette("grey1")
get_palette("grey2")
get_palette("grey3")
get_palette("grey4")
get_palette("grey5")
get_palette("grey6")
get_palette("grey7")
get_palette("grey8")
}
